amba-clcd: Remove check for 'reg' value in clcdfb_probe
This check was attempting to ensure only one clcd device in the
device-tree was probed, however the check fails in the valid case where
the device is a child of another device and the 'reg' value is a offset
from the start of that other device, not an absolute address. This
occurs on vexpress with the motherboard clcd being a child of iofga.

For now, we will just have to rely on there only being one display
device specified in device-tree.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
(cherry picked from commit efff95fcd957f10fc69d8c04a2cdbb07b5c75e40)
Signed-off-by: Alex Shi <alex.shi@linaro.org>